@@ -11,99 +11,75 @@
     "start": "npm run dev",
     "upgrade": "react-native-git-upgrade && npm install --save react-native",
     "upgrade:major": "react-native upgrade",
-    "test": "jest -u",
+    "test": "jest",
     "start:packager": "node_modules/react-native/packager/packager.sh",
     "start:reset": "node node_modules/react-native/local-cli/cli.js start --reset-cache",
     "android": "node node_modules/react-native/local-cli/cli.js run-android",
+    "android:shake": "adb shell input keyevent 82",
+    "android:log": "adb logcat '*:S' ReactNative:V ReactNativeJS:V",
+    "android:apk": "cd android && ./gradlew assembleRelease",
+    "android:release": "node node_modules/react-native/local-cli/cli.js run-android --variant=release",
     "ios": "node node_modules/react-native/local-cli/cli.js run-ios",
+    "ios:release": "node node_modules/react-native/local-cli/cli.js run-ios --configuration Release",
+    "ios:devices": "xcrun simctl list devices",
+    "ios:dooboolab": "node node_modules/react-native/local-cli/cli.js run-ios --device 'dooboolab'",
+    "ios:dooboolab-release": "node node_modules/react-native/local-cli/cli.js run-ios --configuration Release --device 'dooboolab'",
+    "ios:iphone5": "node node_modules/react-native/local-cli/cli.js run-ios  --simulator='iPhone 5s'",
+    "ios:iphone7": "node node_modules/react-native/local-cli/cli.js run-ios  --simulator='iPhone 7'",
     "ios:reset": "rm -rf ios/build/ModuleCache",
-    "jest": "jest",
-    "lint": "eslint -c .eslintrc.js './src/**/*.js'",
-    "build:ios": "react-native bundle --entry-file index.js --platform ios --dev false --bundle-output ios/main.jsbundle --assets-dest ios",
-    "build:android": "react-native bundle --entry-file index.js --platform android --dev false --bundle-output android/main.jsbundle --assets-dest android",
-    "reload:android": "adb shell input keyevent 82"
+    "lint": "eslint src --ext .ts,.tsx,.js,.jsx",
+    "build:ios": "react-native bundle --entry-file index.js --platform ios --dev false --bundle-output ios/main.jsbundle --assets-dest ./ios",
+    "build:android": "react-native bundle --entry-file index.js --platform android --dev false --bundle-output android/main.jsbundle --assets-dest ./android"
   },
   "dependencies": {
-    "mobx": "^5.8.0",
-    "mobx-react": "^6.1.2",
-    "moment": "^2.23.0",
-    "react": "^16.8.6",
-    "react-native": "^0.60.0",
-    "react-native-localization": "^2.0.3",
-    "react-native-audio-recorder-player": "file:../",
-    "react-native-status-bar-height": "^2.2.0",
-    "react-navigation": "^4.0.0"
+    "moment": "^2.24.0",
+    "react": "16.8.6",
+    "react-native": "0.60.5",
+    "react-native-audio-recorder-player": "^2.2.4",
+    "react-native-gesture-handler": "git://github.com/hyochan/react-native-gesture-handler.git#package-names",
+    "react-native-localization": "^2.1.6",
+    "react-native-reanimated": "^1.2.0",
+    "react-navigation": "^4.0.2",
+    "react-navigation-stack": "^1.5.4",
+    "styled-components": "^4.3.2"
   },
   "devDependencies": {
-    "babel-jest": "24.5.0",
-    "babel-preset-react-native": "4.0.1",
-    "jest": "24.5.0",
-    "react-test-renderer": "16.8.6",
-    "@types/enzyme": "^3.1.15",
-    "@types/jest": "^24.0.11",
-    "@types/react": "16.8.25",
-    "@types/react-native": "0.60.13",
-    "apsl-react-native-button": "^3.1.1",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.5",
-    "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.7.1",
-    "enzyme-to-json": "^3.3.5",
+    "@babel/cli": "^7.6.0",
+    "@babel/core": "^7.6.0",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-decorators": "7.6.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
+    "@babel/plugin-transform-runtime": "7.6.0",
+    "@babel/preset-typescript": "^7.6.0",
+    "@babel/runtime": "^7.6.0",
+    "@testing-library/react-native": "^4.0.9",
+    "@types/jest": "^24.0.18",
+    "@types/react": "^16.8.25",
+    "@types/react-native": "^0.60.13",
+    "@types/react-navigation": "^3.0.8",
+    "@types/react-test-renderer": "^16.8.6",
+    "@types/styled-components": "^4.1.19",
+    "@typescript-eslint/eslint-plugin": "^2.2.0",
+    "@typescript-eslint/parser": "^2.2.0",
+    "babel-jest": "24.9.0",
+    "babel-preset-react-native": "^5.0.2",
+    "codecov": "^3.5.0",
+    "eslint": "^6.3.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-standard": "^4.0.1",
+    "eslint-plugin-typescript": "^0.14.0",
+    "jest": "^24.9.0",
+    "jsc-android": "^241213.1.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "prettier": "^1.18.2",
     "react-dom": "^16.8.6",
-    "react-native-typescript-transformer": "^1.2.11",
-    "ts-jest": "^24.0.0",
-    "tslib": "^1.9.3",
-    "tslint": "^5.12.0",
-    "tslint-react": "^4.0.0",
-    "typescript": "^3.2.2"
-  },
-  "jest": {
-    "preset": "react-native",
-    "transformIgnorePatterns": [
-      "node_modules/(?!(.*-)?react-(.*-)?native(-.*)?)"
-    ],
-    "setupFiles": [
-      "./test/shimSetup.js",
-      "./test/jestSetup.ts"
-    ],
-    "globals": {
-      "ts-jest": {
-        "useBabelrc": true,
-        "tsConfigFile": "tsconfig.json"
-      }
-    },
-    "transform": {
-      "^.+\\.jsx?$": "<rootDir>/node_modules/babel-jest",
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "ios.ts",
-      "ios.tsx",
-      "android.ts",
-      "android.tsx"
-    ],
-    "moduleNameMapper": {
-      "^@apis(.*)$": "<rootDir>/src/apis$1",
-      "^@screen(.*)$": "<rootDir>/src/screen$1",
-      "^@navigation(.*)$": "<rootDir>/src/navigation$1",
-      "^@models(.*)$": "<rootDir>/src/models$1",
-      "^@stores(.*)$": "<rootDir>/src/stores$1",
-      "^@utils(.*)$": "<rootDir>/src/utils$1",
-      "^@STRINGS(.*)$": "<rootDir>/src/STRINGS$1",
-      "^@assets(.*)$": "<rootDir>/src/assets$1",
-      "^@shared(.*)$": "<rootDir>/src/shared$1",
-      "^@app(.*)$": "<rootDir>/src/app$1",
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/assetsTransformer.js",
-      "\\.(css|less)$": "<rootDir>/test/assetsTransformer.js"
-    },
-    "snapshotSerializers": [
-      "./node_modules/enzyme-to-json/serializer"
-    ]
+    "react-test-renderer": "^16.8.6",
+    "ts-jest": "^24.0.2",
+    "typescript": "^3.6.3"
   }
 }
