@@ -1,5 +1,5 @@
 import { AppRegistry, YellowBox } from 'react-native';
-import App from './src/';
+import App from './src/App';
 
 /**
  * React Native 0.54 warning message ignore.
@@ -7,6 +7,7 @@ import App from './src/';
 YellowBox.ignoreWarnings([
   'Warning: componentWillMount is deprecated',
   'Warning: componentWillReceiveProps is deprecated',
+  'Module RCTImageLoader',
 ]);
 
-AppRegistry.registerComponent('dooboo', () => App);
+AppRegistry.registerComponent('RNAudioRecorderPlayerEx', () => App);
